refactor(rust): extract terminal.rs with PTY management commands
